requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
pdfplumber==0.10.0
openpyxl==3.1.2
google-cloud-documentai==2.20.2
google-auth==2.23.0
google-auth-httplib2==0.1.1
